refactor: router and schema structures
