+++ app/api/products/inventory_table/schema.py
@@ -1,0 +1,14 @@
+from typing import List
+from pydantic import BaseModel
+
+
+class InventoryTableParams(BaseModel):
+    data: str
+    date_time: str
+    elements: List
+    station: str = None
+    year: str = None
+    month: str = None
+    day: str = None
+    missing_indicator: str = "M"
+    observed_indicator: str = "X"
